TMC5130_eval: Added module IDs to all functions
The IC functions take a channel argument and pass it to the read/write
register functions. For the IC the actual function of that argument is
irrelevant. For the implementation of the Eval class this number is used
for the TMCL communication, where a module ID is used to identify the
module to communicate with.

As a result the Eval class now has the name changed from "channel" to
"moduleID" and now also receives it as a constructor parameter.