tmcl_interface
Added data_available method.
It indicates if and how much data is available to read on the interface.
This might return a boolean (if data is available) or an int (how much data is available) depending on the interface.
Implementation is generally optional, but required if non-blocking loops are required.